added excel basic parsing

+++ GenericBackend/GenericBackend.DataModels/Actual/ActualTimelineData.cs
@@ -6,10 +6,10 @@ namespace GenericBackend.DataModels.Actual
     public class ActualTimelineData : MongoEntityBase
     {
         public DateTime DateTime { get; set; }
-        public long Actual { get; set; }
-        public long UpdateActual { get; set; }
-        public long AccumulatedActual { get; set; }
-        public long AccumulatedUpdate { get; set; }
+        public decimal Actual { get; set; }
+        public decimal UpdateActual { get; set; }
+        public decimal AccumulatedActual { get; set; }
+        public decimal AccumulatedUpdate { get; set; }
         public string SupervisorComments { get; set; }
     }
 }
